# sdk/spec/plugin-api-v1alpha1.openapi.yaml
openapi: 3.1.0
info:
  title: FluidBG Plugin API
  version: 0.2.0
  description: |
    Language-neutral HTTP contract used between the FluidBG operator, plugins,
    and test containers. The plugin API version is fluidbg.plugin/v1alpha1.
jsonSchemaDialect: https://json-schema.org/draft/2020-12/schema
paths:
  /manager/prepare:
    post:
      operationId: managerPrepare
      description: Privileged manager setup for one signed inception point.
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginManagerLifecycleRequest"
      responses:
        "200":
          description: Manager setup completed.
  /manager/cleanup:
    post:
      operationId: managerCleanup
      description: Privileged manager cleanup for one signed inception point.
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginManagerLifecycleRequest"
      responses:
        "200":
          description: Manager cleanup completed.
  /manager/sync:
    post:
      operationId: managerSync
      description: Authenticated manager inventory sync. Managers use this to garbage-collect scoped credentials and plugin-owned resources that missed normal cleanup.
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginManagerSyncRequest"
      responses:
        "200":
          description: Manager sync completed.
  /prepare:
    post:
      operationId: prepare
      description: Non-traffic setup and assignment discovery. Must not start consuming, proxying, observing, writing, or combining traffic.
      security:
        - pluginAuth: []
      responses:
        "200":
          description: Property assignments to apply before activation. Runtime prepare responses may target green and blue only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginLifecycleResponse"
  /activate:
    post:
      operationId: activate
      description: Start traffic work after candidate, verifier, and application rollouts are ready. Must not return assignments.
      security:
        - pluginAuth: []
      responses:
        "200":
          description: Inceptor activated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginLifecycleResponse"
  /drain:
    post:
      operationId: drain
      security:
        - pluginAuth: []
      responses:
        "200":
          description: Property assignments to apply before drain starts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginLifecycleResponse"
  /drain-status:
    get:
      operationId: drainStatus
      security:
        - pluginAuth: []
      responses:
        "200":
          description: Drain status for buffered plugin resources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginDrainStatusResponse"
  /cleanup:
    post:
      operationId: cleanup
      security:
        - pluginAuth: []
      responses:
        "200":
          description: Property assignments to apply after promotion or rollback.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginLifecycleResponse"
  /traffic:
    post:
      operationId: trafficShift
      description: Live progressive traffic update. Must not require a plugin restart.
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficShiftRequest"
      responses:
        "200":
          description: Updated traffic percentage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficShiftResponse"
  /testcases:
    post:
      operationId: registerTestCase
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTestCaseRequest"
      responses:
        "201":
          description: Test case registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterTestCaseResponse"
  /observe/{testId}/{inceptionPoint}:
    post:
      operationId: observe
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
        - name: inceptionPoint
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationNotification"
      responses:
        "200":
          description: Observation accepted.
  /write:
    post:
      operationId: write
      description: Optional plugin writer endpoint used by test containers to inject traffic.
      security:
        - pluginAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HttpWriteRequest"
      responses:
        "200":
          description: Write accepted or forwarded.
components:
  securitySchemes:
    pluginAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Per-inception JWT signed by the operator and passed in Authorization.
  schemas:
    PluginApiVersion:
      type: string
      const: fluidbg.plugin/v1alpha1
    CrdGroup:
      type: string
      const: fluidbg.io
    CrdVersion:
      type: string
      const: v1alpha1
    PluginRole:
      type: string
      enum: [duplicator, splitter, combiner, observer, mock, writer, consumer]
    TrafficRoute:
      type: string
      enum: [blue, green, both, unknown]
    TestIdSelector:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
        jsonPath:
          type: string
        pathSegment:
          type: integer
        value:
          type: string
    FilterCondition:
      type: object
      required: [field]
      additionalProperties: false
      properties:
        field:
          type: string
        equals:
          type: string
        matches:
          type: string
        jsonPath:
          type: string
    ObserverConfig:
      type: object
      additionalProperties: false
      properties:
        testId:
          $ref: "#/components/schemas/TestIdSelector"
        match:
          type: array
          items:
            $ref: "#/components/schemas/FilterCondition"
        notifyPath:
          type: string
    PayloadSelection:
      type: string
      enum: [request, response, both]
    NotificationFilter:
      type: object
      additionalProperties: false
      properties:
        match:
          type: array
          items:
            $ref: "#/components/schemas/FilterCondition"
          default: []
        notifyPath:
          type: string
        payload:
          $ref: "#/components/schemas/PayloadSelection"
    AssignmentTarget:
      type: string
      enum: [green, blue, test]
    AssignmentKind:
      type: string
      enum: [env]
    PropertyAssignment:
      type: object
      required: [target, kind, name, value]
      additionalProperties: false
      properties:
        target:
          $ref: "#/components/schemas/AssignmentTarget"
        kind:
          $ref: "#/components/schemas/AssignmentKind"
        name:
          type: string
        value:
          type: string
        containerName:
          type: string
    PluginLifecycleResponse:
      type: object
      additionalProperties: false
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/PropertyAssignment"
          default: []
        inceptorEnv:
          type: array
          description: Environment variables the operator must inject into the per-inception inceptor before the inceptor pod is created.
          items:
            $ref: "#/components/schemas/InceptorEnvVar"
          default: []
        config:
          description: Optional effective inceptor config. Managers use this to return a secured, derived runtime config while keeping transport-specific derivation out of the operator.
    InceptorEnvVar:
      type: object
      required: [name, value]
      additionalProperties: false
      properties:
        name:
          type: string
        value:
          type: string
    PluginManagerLifecycleRequest:
      type: object
      required: [namespace, blueGreenRef, inceptionPoint, plugin, roles, config]
      additionalProperties: false
      properties:
        namespace:
          type: string
        blueGreenRef:
          type: string
        blueGreenUid:
          type: string
        inceptionPoint:
          type: string
        plugin:
          type: string
        roles:
          type: array
          items:
            type: string
        config: {}
        activeInceptions:
          type: array
          items:
            $ref: "#/components/schemas/ActiveInception"
          default: []
    PluginManagerSyncRequest:
      type: object
      required: [plugin]
      additionalProperties: false
      properties:
        plugin:
          type: string
        activeInceptions:
          type: array
          items:
            $ref: "#/components/schemas/ActiveInception"
          default: []
    ActiveInception:
      type: object
      required: [namespace, blueGreenRef, inceptionPoint, plugin, roles, config]
      additionalProperties: false
      properties:
        namespace:
          type: string
        blueGreenRef:
          type: string
        blueGreenUid:
          type: string
        inceptionPoint:
          type: string
        plugin:
          type: string
        roles:
          type: array
          items:
            type: string
        config: {}
    PluginDrainStatusResponse:
      type: object
      required: [drained]
      additionalProperties: false
      properties:
        drained:
          type: boolean
        message:
          type: string
    TrafficShiftRequest:
      type: object
      required: [trafficPercent]
      additionalProperties: false
      properties:
        trafficPercent:
          type: integer
          minimum: 0
          maximum: 100
    TrafficShiftResponse:
      type: object
      required: [trafficPercent]
      additionalProperties: false
      properties:
        trafficPercent:
          type: integer
          minimum: 0
          maximum: 100
    ObservationNotification:
      type: object
      required: [testId, inceptionPoint, route, payload]
      additionalProperties: true
      properties:
        testId:
          type: string
        inceptionPoint:
          type: string
        route:
          $ref: "#/components/schemas/TrafficRoute"
        payload: {}
    RegisterTestCaseRequest:
      type: object
      required: [test_id, blue_green_ref, inception_point]
      additionalProperties: false
      properties:
        test_id:
          type: string
        blue_green_ref:
          type: string
        inception_point:
          type: string
        triggered_at:
          type: string
          format: date-time
        timeout_seconds:
          type: integer
          minimum: 1
        verify_url:
          type: string
    RegisterTestCaseResponse:
      type: object
      required: [test_id, status]
      additionalProperties: false
      properties:
        test_id:
          type: string
        status:
          type: string
    HttpWriteRequest:
      type: object
      required: [payload]
      additionalProperties: false
      properties:
        testId:
          type: string
        payload: {}
        method:
          type: string
        headers:
          type: object
          additionalProperties: true
        path:
          type: string